Feature/OGB-61

* Merge remote-tracking branch 'origin/feature/OGB-61' into feature/OGB-61

* OGB-61
Delete some checks

* Merge remote-tracking branch 'origin/feature/OGB-61' into feature/OGB-61

# Conflicts:
#	Backend/Services/OneGate.Backend.Services.TimeseriesService/ExecutorService.cs

* OGB-61
Struct to class

* OGB-61
Indexing

* OGB-61
Implement EngineType

* OGB-61
Implement EngineType better

* OGB-61
Engine mapping

* OGB-61
Re-try policies

* OGB-61
Count fix

Approved-by: Alexander Krymov

+++ Backend/OneGate.Backend.Database/Models/OhlcTimeseries.cs
@@ -14,18 +14,18 @@ namespace OneGate.Backend.Database.Models
         
         [Required]
         [Column("low")]
-        public float Low { get; set; }
+        public double Low { get; set; }
 
         [Required] 
         [Column("high")]
-        public float High { get; set; }
+        public double High { get; set; }
 
         [Required]
         [Column("open")]
-        public float Open { get; set; }
+        public double Open { get; set; }
 
         [Required]
         [Column("close")]
-        public float Close { get; set; }
+        public double Close { get; set; }
     }
 }
